#859 move ote1 file path closer to use
