+++ src/a12_hub_toolbox/test/_util/a12_str.py
@@ -5,10 +5,6 @@ def belief_mstr_dir_str() -> str:
     return "belief_mstr_dir"
 
 
-def belief_ote1_agg_str() -> Literal["belief_ote1_agg"]:
-    return "belief_ote1_agg"
-
-
 def gut_str() -> Literal["gut"]:
     return "gut"
