+++ src/a18_etl_toolbox/a18_path.py
@@ -1,7 +1,23 @@
 from src.a00_data_toolbox.file_toolbox import create_path
-from src.a01_term_logic.term import BelieverName
+from src.a01_term_logic.term import BelieverName, LabelTerm
 
 STANCE0001_FILENAME = "stance0001.xlsx"
+BELIEF_OTE1_AGG_CSV_FILENAME = "belief_ote1_agg.csv"
+BELIEF_OTE1_AGG_JSON_FILENAME = "belief_ote1_agg.json"
+
+
+def create_belief_ote1_csv_path(belief_mstr_dir: str, belief_label: LabelTerm):
+    """Returns path: belief_mstr_dir\\beliefs\\belief_label\\belief_ote1_agg.csv"""
+    beliefs_dir = create_path(belief_mstr_dir, "beliefs")
+    belief_path = create_path(beliefs_dir, belief_label)
+    return create_path(belief_path, "belief_ote1_agg.csv")
+
+
+def create_belief_ote1_json_path(belief_mstr_dir: str, belief_label: LabelTerm):
+    """Returns path: belief_mstr_dir\\beliefs\\belief_label\\belief_ote1_agg.json"""
+    beliefs_dir = create_path(belief_mstr_dir, "beliefs")
+    belief_path = create_path(beliefs_dir, belief_label)
+    return create_path(belief_path, "belief_ote1_agg.json")
 
 
 def create_stances_dir_path(belief_mstr_dir: str) -> str:
