+++ src/a18_etl_toolbox/test/test_tran/test_belief_believer_time_agg_csv_to_json.py
@@ -6,11 +6,11 @@ from src.a11_bud_logic.test._util.a11_str import (
     believer_name_str,
     bud_time_str,
 )
-from src.a12_hub_toolbox.a12_path import (
+from src.a17_idea_logic.test._util.a17_str import error_message_str
+from src.a18_etl_toolbox.a18_path import (
     create_belief_ote1_csv_path,
     create_belief_ote1_json_path,
 )
-from src.a17_idea_logic.test._util.a17_str import error_message_str
 from src.a18_etl_toolbox.test._util.a18_env import (
     env_dir_setup_cleanup,
     get_module_temp_dir,
